extract the rest for testing
