+++ src/cloaca/api/get_lifers_by_location.py
@@ -1,0 +1,9 @@
+from cloaca.types import get_lifers_from_cache, group_lifers_by_location
+
+
+def get_lifers_by_location(latitude: float, longitude: float, file_id: str):
+    lifers_from_csv = get_lifers_from_cache(file_id)
+
+    lifers_by_location = group_lifers_by_location(lifers_from_csv)
+
+    return lifers_by_location
