+++ src/cloaca/main.py
@@ -1,33 +1,18 @@
 import time
-from uuid import uuid4
 
+from cloaca.api.get_lifers_by_location import get_lifers_by_location
+from cloaca.api.get_nearby_observations import get_nearby_observations
 from cloaca.api.get_new_lifers_by_region import (
     get_filtered_lifers_for_region,
 )
-from cloaca.parsing.parse_ebird_personal_export import parse_csv_from_file_to_lifers
-from cloaca.parsing.parsing_helpers import Lifer
-from cloaca.types import (
-    get_lifers_from_cache,
-    group_lifers_by_location,
-    set_lifers_to_cache,
-)
 
-import os
-from phoebe_bird import AsyncPhoebe
+from cloaca.api.upload_lifers_csv import upload_lifers_csv
+from cloaca.parsing.parsing_helpers import Lifer
+
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi import FastAPI, Request, UploadFile
 
-from dotenv import load_dotenv
-
-load_dotenv()
-INITIAL_CENTER = {"lng": -74.0242, "lat": 40.6941}
-
-
 Cloaca_App = FastAPI()
-
-phoebe_client = AsyncPhoebe(
-    api_key=os.environ.get("EBIRD_API_KEY"),
-)
 
 Cloaca_App.add_middleware(
     CORSMiddleware,
@@ -56,32 +41,18 @@ def health_check():
 
 
 @Cloaca_App.get("/v1/nearby_observations")
-async def get_nearby_observations(latitude: float, longitude: float, file_id: str):
+async def get_nearby_observations_api(latitude: float, longitude: float, file_id: str):
     return get_nearby_observations(latitude, longitude, file_id)
 
 
 @Cloaca_App.get("/v1/lifers_by_location")
-def get_lifers(latitude: float, longitude: float, file_id: str):
-    lifers_from_csv = get_lifers_from_cache(file_id)
-
-    lifers_by_location = group_lifers_by_location(lifers_from_csv)
-
-    return lifers_by_location
+def get_lifers_by_location_api(latitude: float, longitude: float, file_id: str):
+    return get_lifers_by_location(latitude, longitude, file_id)
 
 
 @Cloaca_App.post("/v1/upload_lifers_csv")
-def upload_lifers_csv(file: UploadFile):
-    print("Uploading file", file.filename)
-
-    # turn the file into a pandas dataframe
-    csv = parse_csv_from_file_to_lifers(file)
-    uuid4_str = str(uuid4())
-
-    set_lifers_to_cache(uuid4_str, csv)
-
-    print(f"parsed csv with key {uuid4_str} and length {len(csv)}")
-
-    return {"key": uuid4_str}
+def upload_lifers_csv_api(file: UploadFile):
+    return upload_lifers_csv(file)
 
 
 @Cloaca_App.get("/v1/regional_new_potential_lifers")
